feed/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11
django-autoslug==1.9.3
django-haystack==2.6.1
django-haystack-elasticsearch==0.1.0
elasticsearch==5.0.1
feedparser==5.2.1
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
schedule==0.4.3
six==1.10.0
social-auth-core==1.5.0
urllib3==1.22
